7/move clipboard fun to separate file, replace about with info
